Renamed to babel.utils.
