+++ src/main/scala/org/renci/babel/utils/Comparer.scala
@@ -1,8 +1,8 @@
-package org.renci.babel.validator
+package org.renci.babel.utils
 
 import com.typesafe.scalalogging.LazyLogging
 import org.renci.babel.utils.MemoryUtils
-import org.renci.babel.validator.model.Compendium
+import org.renci.babel.utils.model.Compendium
 import zio.blocking.Blocking
 import zio.stream.ZStream
 import zio.{Chunk, ZIO}
