+++ src/test/scala/org/renci/babel/utils/CLISpec.scala
@@ -1,4 +1,4 @@
-package org.renci.babel.validator
+package org.renci.babel.utils
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
